+++ src/factory/AbstractFactory.java
@@ -1,0 +1,11 @@
+package factory;
+
+import model.Usuario;
+
+public interface AbstractFactory {
+
+	public static Usuario crearUsuarioFactory() {
+		return UsuarioFactory.crearUsuario();
+	}
+
+}
