+++ src/model/Usuario.java
@@ -10,9 +10,6 @@ package model;
 
 import java.time.LocalDate;
 
-import utilidades.Escritor;
-import utilidades.Lector;
-
 public class Usuario {
 
 	private String nombre;
@@ -56,18 +53,6 @@ public class Usuario {
 	public String toString() {
 		return "Usuario [nombre=" + nombre + ", fechaNacimiento=" + fechaNacimiento + ", ciudadResidencia="
 				+ ciudadResidencia + "]";
-	}
-
-	public static Usuario crearUsuario() {
-		Usuario usuario = new Usuario();
-		Escritor.str("vamos a crear un nuevo usuario, voy a pedirte sus datos");
-		usuario.setNombre(Lector.str("dame nombre"));
-		int year = Lector.pedirInt("dame año de nacimiento");
-		int mes = Lector.pedirInt("dame mes de nacimiento");
-		int dia = Lector.pedirInt("dame día de nacimiento");
-		usuario.setFechaNacimiento(LocalDate.of(year, mes, dia));
-		usuario.setCiudadResidencia(Lector.str("dame ciudad de residencia"));
-		return usuario;
 	}
 
 }
